LayoutTests/platform/ios-sim-deprecated/storage/indexeddb/clone-exception-expected.txt

CONSOLE MESSAGE: line 18: TypeError: undefined is not an object (evaluating 'request')
Ensure DataCloneError is consistently thrown by IndexedDB methods

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


dbname = "clone-exception.html"

doFirstOpen():
indexedDB.open(dbname + '1')
FAIL indexedDB.open(dbname + '1') threw exception TypeError: null is not an object (evaluating 'indexedDB.open')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
